Refactor hard-coded modifiers to be applied identically to user-defined modifiers
